requirements.txt
PyJWT
google-auth
requests
Werkzeug<0.13.0,>=0.12.0
Flask==0.12.2
requests-toolbelt==0.7.1
boto3==1.4.7

# Only if using flex
gunicorn==19.9.0